Privacy Risk Score Calculator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
lxml==5.2.1
pyyaml==6.0.2
rich==13.7.1
typer==0.12.3
pytest==8.2.0
